Fix py34 DB issue and remove py33
To solve the problem of "db type could
not be determined" on py34 we have to run first the py34 env to, then,
run py27. This patch puts py34 first on the tox.ini list of envs to
avoid this problem to happen. py33 is not supported from Mitaka, as
per Infra, therefore removing the support for the same.

Closes-Bug: #1526170
Closes-bug: #1489059
Change-Id: I31588ecadcf6f219ad3b31598df7d39fb8c3f695